build warning fix

+++ daw_hash_table.h
@@ -198,7 +198,7 @@ namespace daw {
 		static size_t scale_hash( size_t hash, size_t table_size ) {
 			// Scale value to capacity using MAD(Multiply-Add-Divide) compression
 			// Use the two largest Prime's that fit in a 64bit unsigned integral
-			assert( table_size < std::numeric_limits<ptrdiff_t>::max( ) );	// Table size must be less than max of ptrdiff_t as we use the value 0 as a sentinel.  This should be rare
+			assert( table_size < static_cast<size_t>( std::numeric_limits<ptrdiff_t>::max( )) );	// Table size must be less than max of ptrdiff_t as we use the value 0 as a sentinel.  This should be rare
 			assert( hash != 0 );	// zero is a sentinel for no value
 			static const size_t prime_a = 18446744073709551557u;
 			static const size_t prime_b = 18446744073709551533u;
